requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
aiofiles==23.2.1
PyJWT==2.8.0
pyodbc
azure-identity
pandas
msfabricpysdkcore